+++ tests/Etymon.Base.Tests/Main.fs
@@ -1,0 +1,10 @@
+module Etymon.Base.Tests.Main
+
+open Expecto
+
+[<EntryPoint>]
+let main argv =
+    runTestsWithCLIArgs
+        []
+        argv
+        (testList "Etymon.Base" [ ParseTests.tests; StrTests.tests; DictEnvTests.tests; FilesTests.tests ])